server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["karthik1881 <“Karthik”@itero.ie>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
langchain = "^0.1.16"
langchain-core = "^0.1.46"
langchain-community = "^0.0.34"
python-multipart = "^0.0.9"
unstructured = { version = "^0.12.0", extras = ["pdf"], python = ">=3.11,<3.12" }
langchain-openai = "^0.1.4"
langchain_pinecone = { version = "^0.1.0", python = ">=3.11,<3.12" }
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "run:main"


